+++ docs/examples/metrics/example.py
@@ -5,6 +5,7 @@ from opentelemetry._metrics import (
     get_meter_provider,
     set_meter_provider,
 )
+from opentelemetry._metrics._internal.instrument import CallbackOptions
 from opentelemetry.exporter.otlp.proto.grpc._metric_exporter import (
     OTLPMetricExporter,
 )
@@ -17,15 +18,17 @@ provider = MeterProvider(metric_readers=[reader])
 set_meter_provider(provider)
 
 
-def observable_counter_func() -> Iterable[Observation]:
+def observable_counter_func(options: CallbackOptions) -> Iterable[Observation]:
     yield Observation(1, {})
 
 
-def observable_up_down_counter_func() -> Iterable[Observation]:
+def observable_up_down_counter_func(
+    options: CallbackOptions,
+) -> Iterable[Observation]:
     yield Observation(-10, {})
 
 
-def observable_gauge_func() -> Iterable[Observation]:
+def observable_gauge_func(options: CallbackOptions) -> Iterable[Observation]:
     yield Observation(9, {})
 
 
@@ -37,7 +40,8 @@ counter.add(1)
 
 # Async Counter
 observable_counter = meter.create_observable_counter(
-    "observable_counter", observable_counter_func
+    "observable_counter",
+    [observable_counter_func],
 )
 
 # UpDownCounter
@@ -47,7 +51,7 @@ updown_counter.add(-5)
 
 # Async UpDownCounter
 observable_updown_counter = meter.create_observable_up_down_counter(
-    "observable_updown_counter", observable_up_down_counter_func
+    "observable_updown_counter", [observable_up_down_counter_func]
 )
 
 # Histogram
@@ -55,4 +59,4 @@ histogram = meter.create_histogram("histogram")
 histogram.record(99.9)
 
 # Async Gauge
-gauge = meter.create_observable_gauge("gauge", observable_gauge_func)
+gauge = meter.create_observable_gauge("gauge", [observable_gauge_func])